+++ NEWS.md
@@ -1,5 +1,9 @@
 # rtrek 0.2.0
 
+* Added an initial version of an API function `memory_alpha` along with a collection of internal support functions for accessing Star Trek data from web portals available on the Memory Alpha website.
+    * Internal portal functions that return data from portal and portal category web pages, and which may run recursively when necessary to compile complete content over multiple pages, are memoized. 
+    * Functions like `ma_article`, which return article content from a terminal endpoint are not memoized.
+* Added `ma_search` function.
 * Added new datasets, mainly to timelines: `tlBooks`, `tlEvents`, `tlFootnotes`. Also added `stSeries` names and abbreviations table.
 * Renamed `st_book_series` and `stBooksWP` to `st_books_wiki` and `stBooksWiki`, respectively.
 * Updated tile set URLs.