+++ NEWS.md
@@ -32,7 +32,7 @@
 
 * Remove remote `trekfont` now that it is available on CRAN.
 * Prepare GitHub accessible external datasets that serve package web pages.
-* Updated documentation, webste. Add to "fun pages" (R & R), which are only part of website "vignettes", not package vignettes.
+* Updated documentation, website. Add to "fun pages" (R & R), which are only part of website "vignettes", not package vignettes.
 * Minor fixes.
 
 # rtrek 0.0.8